Move subcommands to each sub packages
